Implicit casts should not lower precision

+++ src/drm.h
@@ -92,7 +92,7 @@ class Property : public Object {
   bool isImmutable() const { return flags_ & DRM_MODE_PROP_IMMUTABLE; }
 
   std::vector<uint64_t> values() const { return values_; }
-  const std::map<uint32_t, std::string>& enums() const { return enums_; }
+  const std::map<__u64, std::string>& enums() const { return enums_; }
   std::vector<uint32_t> blobs() const { return blobs_; }
 
  private:
@@ -100,7 +100,7 @@ class Property : public Object {
   std::string name_;
   uint32_t flags_;
   std::vector<uint64_t> values_;
-  std::map<uint32_t, std::string> enums_;
+  std::map<__u64, std::string> enums_;
   std::vector<uint32_t> blobs_;
 };
